# api/openapi.yaml
openapi: 3.0.3
info:
  title: openbot API
  version: 0.1.0
  description: |
    HTTP API for openbotd. Authenticated with a session cookie (`openbot_session`).
    Mutating requests must send `X-Openbot-CSRF: 1` (checked together with the Origin header).

    Realtime updates are delivered over a WebSocket at `GET /api/ws`; each frame is a JSON
    `WsEvent` (see components). The WebSocket is not described as an operation because
    OpenAPI can't express it; its payloads are.
servers:
  - url: /api
paths:
...
  /auth/setup:
    post:
      operationId: setupAccount
      tags: [auth]
      security: []
      description: Create the single user account. Only allowed while no account exists.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Credentials" }
      responses:
        "200":
          description: Account created and session started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
  /chats/{chatId}/attachments:
    parameters:
      - $ref: "#/components/parameters/ChatId"
    post:
      operationId: uploadAttachment
      tags: [chats]
      description: |
        Upload a file to attach to a message in this chat (then pass its id in
        SendMessageRequest.attachmentIds). Up to 25 MB. Uploads that are never sent are removed
        after a day.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: The uploaded file
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "413":
          description: The file is over 25 MB
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /attachments/{attachmentId}:
    parameters:
      - name: attachmentId
        in: path
        required: true
        schema: { type: string }
      - name: download
        in: query
        required: false
        description: Send as a download even if the browser could show it
        schema: { type: boolean }
    get:
      operationId: getAttachment
      tags: [chats]
      description: |
        The file's contents. Images (PNG, JPEG, GIF, WebP) and PDFs are shown in the browser;
        everything else downloads.
      responses:
        "200":
          description: The file
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: openbot_session
  parameters:
    ChatId:
      name: chatId
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code:
          type: string
          description: |
            Machine-readable reason where clients react to it: sign_in_required (an MCP server
            wants a sign-in instead of a key), reconnect (a sign-in expired)
...
    Credentials:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 64 }
        password: { type: string, minLength: 12, maxLength: 256 }
    User:
      type: object
      required: [id, username]
      properties:
        id: { type: string }
        username: { type: string }
...
    Attachment:
      type: object
      required: [id, name, mime, size, url, width, height]
      properties:
        id: { type: string }
        name: { type: string, description: The file name }
        mime: { type: string, description: 'e.g. "image/png", "application/pdf"' }
        size: { type: integer, format: int64, description: Bytes }
        url: { type: string, description: Where to fetch it (same origin, needs the session) }
        width: { type: integer, nullable: true, description: Pixels, for images }
        height: { type: integer, nullable: true, description: Pixels, for images }
...
security:
  - session: []